+++ packages/storage_client/lib/src/file_stub.dart
@@ -1,0 +1,2 @@
+/// A stub for the `dart:io` [File] class.
+typedef File = dynamic;